hotelmanagement/src/test/java/com/nuance/him/config/roomconfig/RoomServiceConfigTest.java
/*
 * COPYRIGHT: Copyright (c) 2018 by Nuance Communications, Inc.
 *  Warning: This product is protected by United States copyright law. Unauthorized use or duplication of this software, in whole or in part, is prohibited.
 *
 */

package com.nuance.him.config.roomconfig;

import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
import org.testng.annotations.Test;
import com.nuance.him.service.RoomService;
import static org.testng.Assert.assertNotNull;
@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes={ RoomServiceConfig.class})
public class RoomServiceConfigTest extends AbstractTestNGSpringContextTests {

   @Autowired
    private RoomService roomService;


    @Test
    public void testRoomService(){
        assertNotNull(roomService);

    }
}